Fix integer literal size bug & add automatic work-group padding

+++ gpgpu-sim-complete/src/cuda-sim/ptx_ir.h
@@ -722,7 +722,8 @@ public:
    {
       ptx_reg_t result;
       switch ( m_type ) {
-      case int_t:         result.s32 = m_value.m_int; break;
+      // deicide: PTX uses 64 bit for signed integer literal
+      case int_t:         result.s64 = m_value.m_int; break;
       case float_op_t:    result.f32 = m_value.m_float; break;
       case double_op_t:   result.f64 = m_value.m_double; break; 
       case unsigned_t:    result.u32 = m_value.m_unsigned; break;
